pyproject.toml
[project]
name = "wechat-artical"
version = "0.1.0"
description = "WeChat content reconstruction and draft pipeline"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "alembic>=1.14.1,<2.0.0",
  "beautifulsoup4>=4.13.3,<5.0.0",
  "fastapi>=0.115.8,<1.0.0",
  "httpx>=0.28.1,<1.0.0",
  "playwright>=1.52.0,<2.0.0",
  "pydantic-settings>=2.8.1,<3.0.0",
  "psycopg[binary]>=3.2.6,<4.0.0",
  "redis>=5.2.1,<6.0.0",
  "sqlalchemy>=2.0.38,<3.0.0",
  "uvicorn>=0.34.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4,<9.0.0",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
